Add automatic database cleanup and fix bug
The database now stays in a clean state when servers or channels are
added or removed, even when the bot is offline when it happens. This
required major restructuring of the database and a new ServerData class.
This makes great progress toward Issue #2 on the GitHub issues page.

This commit also fixes a bug that made it so that only one user was
tracked for slowmodes at a time in each channel.