Make bot aware of Discord permissions
The bot now only lets users with the correct permissions create or
remove slowmodes. This makes progress towards Issue #5 on the GitHub
issues page.

The bot needs special behavior for admins or when a role
is above another role. A user should not be able to set a slowmode on
themselves or on users higher than them in the hierarchy.

The bot also ignores duplicate excluded or included members and if a
user is both excluded AND included, it cancels the set operation.