Split commands into TypeScript classes
I'm very sorry that I couldn't get git to track the changes to the new
files or new file locations in the case of main.js. The commands are
mostly unmodified outside of adapting them to work in the new structure.

The output of commands has been improved through setting return values.
This is especially the case in the Set command.