Fix Remove command permission checking
It only checked the permissions where the command was being called
*from*, and not where the command was being called *on*.

This also required moving the permission checking code so that each
command manages its own permissions, instead of having main.js check the
permissions for them. I don't think there's any good reason to have it
one way or the other. I just changed it to fix the bug in Remove.