Make Set command more efficient with network requests
This update required changing the way that the Status of a slowmode is
printed when a new slowmode is Set. Previously,
Command.getSlowmodeSubjects would fetch the user and role objects every
time it is called, but this isn't always necessary because Set already
has the objects. Set can just pass them to the method. This has now been
fixed. The Status command had to be changed so that it fetches the
objects instead of letting the method do it.

This update also fixes a bug that allowed users to set a slowmode on
channels that already have a slowmode that specially includes them. This
allowed someone to bypass a -includes if they have the permissions for
it, rendering it useless.