Split up functions in and clean up Set.ts
The private function "command" that creates the slowmode should take
GuildMember and Role objects instead of taking object IDs (Snowflakes)
and converting them before use. This reduces the complexity of the
function and makes the code look cleaner since less typecasting is
required.

This change also fixes a bug that would crash the bot if multiple
users/roles were included/excluded, but one of them was invalid or a
duplicate of another. The bot would then compare the author with an
undefined user since the list shinks in size while iterating through it.