Make Set silently ignore duplicate user/role tags
This builds on the previous commit. Before these updates, duplicate user
tags would crash the bot, and duplicate role tags would go unnoticed and
make duplicate entries in the database. Now, the user is allowed to add
duplicate tags, but they are removed internally. Eventually the DB
should be pruned of the duplicate roles.