Fix bug in Remove command that crashes bot when invalid tag is provided
Attempting to remove an invalid channel tag (e.g "<#53535>") would crash
the bot. Now, the bot checks that that channel exists and is in the
server before continuing.

Part of this fix involved reducing the responsibilities of the private
command() function and adding comments. This needs to be done to more
Command files to ensure that no other bugs are being overlooked.