Convert Database and ChannelData to TypeScript and document them
This commit also fixes a bug with Database#setChannel. Previously, the
function would execute extra SELECT statements to get the row, even when
it wasn't needed to INSERT or UPDATE the row. Now once a row is queried,
it is preserved by main and the Commands, and UPDATE commands are only
executed when variables actually change. This eliminates the unnecessary
database queries, thus speeding up the bot.