Move main.subjectToSlowmode to ChannelData
The function is no longer passed as a parameter to the Commands that
need it. It is an instance command, so it now benefits from typing. Some
casting is now being done where it is used in Remove and Reset, but when
the bot upgrades to Discord.js 14 its new types should allow the casting
to be removed.

A few helper functions in ChannelData were removed since they were just
for the benefit of subjectToSlowmode.